chore!: Define single ecdsa constraint generator (#16612)
Audit part 5: Define single ecdsa constraint generator

In this PR we merge the two ECDSA constraint generators (secp256k1 and
secp256r1) into a single one templated over `Curve` type. We also
restructure the code of the constraint generator to be more similar to
other parts of the `dsl` package.

There are some changes to the circuits. This is due to some
`assert_equal` that have been removed, and `normalize` calls that are
not needed.

We also add documentation and define a `utils.hpp` file in the `dsl`
package containing some functions that are used multiple times in
different files.