chore!: EC add from ACIR, part 1 (#18013)
### 🧾 Audit Context

This PR simply refactors the EC addition constraints. The aim is to
simplify the structure of some functions, and add some documentation.
The change in VK is explained below.

### 🛠️ Changes Made

The change of VKs comes from:
- casting `predicate` to bool_t
- the usage of `assert_equal` from `cycle_group` instead of individually
checking the components; the change is due to the fact that the
`cycle_group::assert_equal` method standardises both inputs, while
previously we were standardising only the result computed (`input1 +
input2`)

These changes make the constraints more restrictive (for the predicate)
and ensure that the constraint are complete (if `input_result` is not
standardised, checking all the components might fail as `result` is
standardised)

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [ ] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [ ] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [ ] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers