chore!: improve poly allocation in ZK setting (#18067)
Prior to this PR for the ZK setting we would allocate all prover
polynomials to the full dyadic size. This was a shortcut since _certain_
polynomials (witnesses) are blinded in the final coeffs and therefore
need to be full sized. This PR updates the logic so that all polynomials
are allocated with the minimum possible size (while accounting for
blinding in witnesses)

Note: also simplifies table poly construction so that tables are
constructed starting at the top of the trace rather than aligned with
the lookup block. This was only really useful in the structured trace
setting.